Experimental: start adapter without rebooting it
For some use-cases, restarting the adapter takes a considerable amount
of time. It's a riskier proposition to start up with unknown adapter
state, but in many cases it's really not that bad. In practice, it can
work really well - interactive command-line applications for pygatt can
open and reload previous connections in 100's of milliseconds.

This is experimental because there are still open issues.

What works well:
* When the adapter's not subscribed to much, works seamlessly
* Good at re-instantiating existing connections on start

What doesn't always work well:
* State management isn't great: in some instances attempts to disable
advertising and bond status fail. But in general ignoring the failures
works OK
* If the adapter is in the middle of receiving a lot of notifications,
it may fail. A crude retry mechanism has been added, which works in a
pinch. But it causes latency and occasionally fails to initialize
altogether.

The main TODO: instead of retry logic, reset state more forcefully.
Either
A) Hammer off any active notification streams before starting reader
thread
OR
B) Patch underlying issues causing reader thread to fail to synchronize
when launching in the midst of a lot of unrelated data.