Execute packet handlers immediately
Currently someone needs to call an "expect" in order for any of the
packet handlers to get called. But if you're just sitting idly waiting
for notifications, you probably aren't calling "expect" an awful lot. So
the reader thread should go ahead and call the packet handlers directly.
This way applications can register to disconnected events and respond
immediately!